Move RtcDateTime to dedicated driver
Without this change, we cannot have RtcDateTime with OsTimer.
