+++ src/lib.rs
@@ -35,6 +35,7 @@ pub mod i2c;
 pub mod iopctl;
 pub mod pwm;
 pub mod rng;
+pub mod rtc;
 /// Time driver for the iMX RT600 series.
 #[cfg(feature = "_time-driver")]
 pub mod time_driver;
